계층별 모듈 분리 UseCase 적용 EventBus 추상화

+++ app/src/main/java/io/github/devy/architecture_study/data/repository/UserRepositoryImpl.kt
@@ -7,7 +7,6 @@ import io.github.devy.architecture_study.data.model.toUser
 import io.github.devy.architecture_study.domain.model.User
 import io.github.devy.architecture_study.domain.repositoty.UserRepository
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.withContext
 
 class UserRepositoryImpl(
@@ -16,8 +15,6 @@ class UserRepositoryImpl(
 ) : UserRepository {
 
     private val userCache = linkedMapOf<Int, User>()
-    private val likeEventBus = LikeEventBus()
-
 
     override suspend fun getUsers(query: String): List<User> = withContext(Dispatchers.IO) {
         val likeUserSet = userDao.getLikeUsers().map { it.userId }.toSet()
@@ -29,10 +26,8 @@ class UserRepositoryImpl(
     }
 
     override suspend fun updateLike(userId: Int, like: Boolean): Boolean {
-        val success = if (like) userDao.insertLikeUser(LikeUserEntity(userId)) > -1
+        return if (like) userDao.insertLikeUser(LikeUserEntity(userId)) > -1
         else userDao.deleteLikeUser(userId) > 0
-        if (success) likeEventBus.produceEvent(userId, like)
-        return success
     }
 
     override fun getUserFromCache(userId: Int) = userCache[userId]
@@ -41,7 +36,5 @@ class UserRepositoryImpl(
         return if (onlyLike) userCache.values.filter { it.like }
         else userCache.values.toList()
     }
-
-    override fun getLikeChangeEvent(): Flow<Pair<Int, Boolean>> = likeEventBus.events
 
 }
